Refactor unpacking/helpers in ewald+pme, extract slab.py, _utils.py
