+++ nvalchemiops/torch/interactions/electrostatics/__init__.py
@@ -28,7 +28,6 @@ from nvalchemiops.torch.interactions.electrostatics.dsf import (
     dsf_coulomb,
 )
 from nvalchemiops.torch.interactions.electrostatics.ewald import (
-    compute_slab_correction,
     ewald_real_space,
     ewald_reciprocal_space,
     ewald_summation,
@@ -51,6 +50,9 @@ from nvalchemiops.torch.interactions.electrostatics.pme import (
     pme_energy_corrections_with_charge_grad,
     pme_green_structure_factor,
     pme_reciprocal_space,
+)
+from nvalchemiops.torch.interactions.electrostatics.slab import (
+    compute_slab_correction,
 )
 
 __all__ = [
