+++ nvalchemiops/torch/interactions/electrostatics/_util.py
@@ -17,9 +17,27 @@
 
 from __future__ import annotations
 
+from typing import NamedTuple
+
 import torch
 
-__all__ = ["_InjectChargeGrad", "_sum_charge_gradients"]
+__all__ = [
+    "ElectrostaticOutputs",
+    "_InjectChargeGrad",
+    "_build_electrostatic_result",
+    "_combine_electrostatic_outputs",
+    "_sum_charge_gradients",
+    "_unpack_electrostatic_outputs",
+]
+
+
+class ElectrostaticOutputs(NamedTuple):
+    """Named electrostatics outputs in public API order."""
+
+    energies: torch.Tensor
+    forces: torch.Tensor | None = None
+    charge_grads: torch.Tensor | None = None
+    virial: torch.Tensor | None = None
 
 
 @torch.compiler.disable
@@ -29,6 +47,131 @@ def _sum_charge_gradients(
 ) -> torch.Tensor:
     """Sum electrostatic charge gradients eagerly on compiled paths."""
     return real_space_charge_grads + reciprocal_charge_grads
+
+
+def _build_electrostatic_result(
+    outputs: ElectrostaticOutputs,
+    compute_forces: bool,
+    compute_charge_gradients: bool,
+    compute_virial: bool,
+) -> torch.Tensor | tuple[torch.Tensor, ...]:
+    """Build an output tuple in electrostatics API order."""
+    result = [outputs.energies]
+    if compute_forces and outputs.forces is not None:
+        result.append(outputs.forces)
+    if compute_charge_gradients and outputs.charge_grads is not None:
+        result.append(outputs.charge_grads)
+    if compute_virial and outputs.virial is not None:
+        result.append(outputs.virial)
+    return tuple(result) if len(result) > 1 else result[0]
+
+
+def _unpack_electrostatic_outputs(
+    outputs: torch.Tensor | tuple[torch.Tensor, ...],
+    compute_forces: bool,
+    compute_charge_gradients: bool,
+    compute_virial: bool,
+) -> ElectrostaticOutputs:
+    """Unpack electrostatics outputs by flag combination without cursor logic."""
+    output_tuple = outputs if isinstance(outputs, tuple) else (outputs,)
+
+    if compute_forces and compute_charge_gradients and compute_virial:
+        energies, forces, charge_grads, virial = output_tuple
+    elif compute_forces and compute_charge_gradients:
+        energies, forces, charge_grads = output_tuple
+        virial = None
+    elif compute_forces and compute_virial:
+        energies, forces, virial = output_tuple
+        charge_grads = None
+    elif compute_charge_gradients and compute_virial:
+        energies, charge_grads, virial = output_tuple
+        forces = None
+    elif compute_forces:
+        energies, forces = output_tuple
+        charge_grads = None
+        virial = None
+    elif compute_charge_gradients:
+        energies, charge_grads = output_tuple
+        forces = None
+        virial = None
+    elif compute_virial:
+        energies, virial = output_tuple
+        forces = None
+        charge_grads = None
+    else:
+        (energies,) = output_tuple
+        forces = None
+        charge_grads = None
+        virial = None
+
+    return ElectrostaticOutputs(energies, forces, charge_grads, virial)
+
+
+def _combine_electrostatic_outputs(
+    real_outputs: torch.Tensor | tuple[torch.Tensor, ...],
+    reciprocal_outputs: torch.Tensor | tuple[torch.Tensor, ...],
+    slab_outputs: torch.Tensor | tuple[torch.Tensor, ...] | None,
+    compute_forces: bool,
+    compute_charge_gradients: bool,
+    compute_virial: bool,
+) -> torch.Tensor | tuple[torch.Tensor, ...]:
+    """Combine real, reciprocal, and optional slab outputs by named fields."""
+    real = _unpack_electrostatic_outputs(
+        real_outputs, compute_forces, compute_charge_gradients, compute_virial
+    )
+    reciprocal = _unpack_electrostatic_outputs(
+        reciprocal_outputs, compute_forces, compute_charge_gradients, compute_virial
+    )
+
+    energies = real.energies + reciprocal.energies
+    forces = (
+        real.forces + reciprocal.forces
+        if compute_forces and real.forces is not None and reciprocal.forces is not None
+        else None
+    )
+
+    if (
+        compute_charge_gradients
+        and real.charge_grads is not None
+        and reciprocal.charge_grads is not None
+    ):
+        if torch.compiler.is_compiling():
+            charge_grads = _sum_charge_gradients(
+                real.charge_grads, reciprocal.charge_grads
+            )
+        else:
+            charge_grads = real.charge_grads + reciprocal.charge_grads
+    else:
+        charge_grads = None
+
+    virial = (
+        real.virial + reciprocal.virial
+        if compute_virial and real.virial is not None and reciprocal.virial is not None
+        else None
+    )
+
+    if slab_outputs is not None:
+        slab = _unpack_electrostatic_outputs(
+            slab_outputs, compute_forces, compute_charge_gradients, compute_virial
+        )
+        energies = energies + slab.energies
+        if compute_forces and forces is not None and slab.forces is not None:
+            forces = forces + slab.forces
+        if (
+            compute_charge_gradients
+            and charge_grads is not None
+            and slab.charge_grads is not None
+        ):
+            charge_grads = charge_grads + slab.charge_grads
+        if compute_virial and virial is not None and slab.virial is not None:
+            virial = virial + slab.virial
+
+    return _build_electrostatic_result(
+        ElectrostaticOutputs(energies, forces, charge_grads, virial),
+        compute_forces,
+        compute_charge_gradients,
+        compute_virial,
+    )
 
 
 class _InjectChargeGrad(torch.autograd.Function):
