+++ nvalchemiops/torch/interactions/electrostatics/ewald.py
@@ -26,7 +26,6 @@ Public API
 - ``ewald_real_space()``: Real-space component of Ewald summation
 - ``ewald_reciprocal_space()``: Reciprocal-space component
 - ``ewald_summation()``: Complete Ewald summation (real + reciprocal)
-- ``compute_slab_correction()``: Standalone 2D slab correction for Ewald outputs
 
 Mathematical Formulation
 ------------------------
@@ -98,10 +97,6 @@ from nvalchemiops.interactions.electrostatics.ewald_kernels import (
     _ewald_reciprocal_space_virial_kernel_overload,
     _ewald_subtract_self_energy_kernel_overload,
 )
-from nvalchemiops.interactions.electrostatics.slab_kernels import (
-    _launch_slab_correction,
-    _slab_reduce_moments_kernel_overload,
-)
 from nvalchemiops.torch.autograd import (
     OutputSpec,
     WarpAutogradContextManager,
@@ -111,8 +106,11 @@ from nvalchemiops.torch.autograd import (
     warp_from_torch,
 )
 from nvalchemiops.torch.interactions.electrostatics._util import (
+    ElectrostaticOutputs,
+    _build_electrostatic_result,
+    _combine_electrostatic_outputs,
     _InjectChargeGrad,
-    _sum_charge_gradients,
+    _unpack_electrostatic_outputs,
 )
 from nvalchemiops.torch.interactions.electrostatics.k_vectors import (
     generate_k_vectors_ewald_summation,
@@ -120,10 +118,12 @@ from nvalchemiops.torch.interactions.electrostatics.k_vectors import (
 from nvalchemiops.torch.interactions.electrostatics.parameters import (
     estimate_ewald_parameters,
 )
+from nvalchemiops.torch.interactions.electrostatics.slab import (
+    compute_slab_correction as _compute_slab_correction,
+)
 from nvalchemiops.torch.types import get_wp_dtype, get_wp_mat_dtype, get_wp_vec_dtype
 
 __all__ = [
-    "compute_slab_correction",
     "ewald_real_space",
     "ewald_reciprocal_space",
     "ewald_summation",
@@ -2648,30 +2648,6 @@ def ewald_real_space(
     # compute_virial=True), we must dispatch a force-capable kernel.
     need_force_kernel = compute_forces or compute_virial
 
-    # Helper to build the return tuple from raw outputs using match dispatch.
-    def _build_result(energies, forces=None, charge_grads=None, virial=None):
-        match (
-            compute_forces and forces is not None,
-            compute_charge_gradients and charge_grads is not None,
-            compute_virial and virial is not None,
-        ):
-            case (True, True, True):
-                return energies, forces, charge_grads, virial
-            case (True, True, False):
-                return energies, forces, charge_grads
-            case (True, False, True):
-                return energies, forces, virial
-            case (True, False, False):
-                return energies, forces
-            case (False, True, True):
-                return energies, charge_grads, virial
-            case (False, True, False):
-                return energies, charge_grads
-            case (False, False, True):
-                return energies, virial
-            case _:
-                return energies
-
     if hybrid_forces:
         pos_d = positions.detach()
         chg_d = charges.detach()
@@ -2745,7 +2721,12 @@ def ewald_real_space(
                 energies, charges, charge_grads, batch_idx
             )
 
-        return _build_result(energies, forces, charge_grads, virial)
+        return _build_electrostatic_result(
+            ElectrostaticOutputs(energies, forces, charge_grads, virial),
+            compute_forces,
+            compute_charge_gradients,
+            compute_virial,
+        )
 
     if compute_charge_gradients:
         if neighbor_list is not None:
@@ -2811,7 +2792,12 @@ def ewald_real_space(
         else:
             raise ValueError("Either neighbor_list or neighbor_matrix must be provided")
 
-        return _build_result(energies, forces, charge_grads, virial)
+        return _build_electrostatic_result(
+            ElectrostaticOutputs(energies, forces, charge_grads, virial),
+            compute_forces,
+            compute_charge_gradients,
+            compute_virial,
+        )
 
     # No charge gradients requested
     if neighbor_list is not None:
@@ -2830,7 +2816,12 @@ def ewald_real_space(
                     neighbor_shifts,
                     compute_virial=compute_virial,
                 )
-                return _build_result(energies, forces, virial=virial)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies, forces, virial=virial),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
             else:
                 energies = _batch_ewald_real_space_energy(
                     positions,
@@ -2842,7 +2833,12 @@ def ewald_real_space(
                     neighbor_ptr,
                     neighbor_shifts,
                 )
-                return _build_result(energies)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
         else:
             if need_force_kernel:
                 energies, forces, virial = _ewald_real_space_energy_forces(
@@ -2855,7 +2851,12 @@ def ewald_real_space(
                     neighbor_shifts,
                     compute_virial=compute_virial,
                 )
-                return _build_result(energies, forces, virial=virial)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies, forces, virial=virial),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
             else:
                 energies = _ewald_real_space_energy(
                     positions,
@@ -2866,7 +2867,12 @@ def ewald_real_space(
                     neighbor_ptr,
                     neighbor_shifts,
                 )
-                return _build_result(energies)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
     elif neighbor_matrix is not None:
         if is_batch:
             if need_force_kernel:
@@ -2881,7 +2887,12 @@ def ewald_real_space(
                     mask_value,
                     compute_virial=compute_virial,
                 )
-                return _build_result(energies, forces, virial=virial)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies, forces, virial=virial),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
             else:
                 energies = _batch_ewald_real_space_energy_matrix(
                     positions,
@@ -2893,7 +2904,12 @@ def ewald_real_space(
                     neighbor_matrix_shifts,
                     mask_value,
                 )
-                return _build_result(energies)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
         else:
             if need_force_kernel:
                 energies, forces, virial = _ewald_real_space_energy_forces_matrix(
@@ -2906,7 +2922,12 @@ def ewald_real_space(
                     mask_value,
                     compute_virial=compute_virial,
                 )
-                return _build_result(energies, forces, virial=virial)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies, forces, virial=virial),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
             else:
                 energies = _ewald_real_space_energy_matrix(
                     positions,
@@ -2917,7 +2938,12 @@ def ewald_real_space(
                     neighbor_matrix_shifts,
                     mask_value,
                 )
-                return _build_result(energies)
+                return _build_electrostatic_result(
+                    ElectrostaticOutputs(energies),
+                    compute_forces,
+                    compute_charge_gradients,
+                    compute_virial,
+                )
     else:
         raise ValueError("Either neighbor_list or neighbor_matrix must be provided")
 
@@ -2991,30 +3017,6 @@ def ewald_reciprocal_space(
     elif not is_batch and k_vectors.dim() == 3 and k_vectors.shape[0] == 1:
         k_vectors = k_vectors.squeeze(0)
 
-    # Helper to build the return tuple from raw outputs using match dispatch.
-    def _build_result(energies, forces=None, charge_grads=None, virial=None):
-        match (
-            compute_forces and forces is not None,
-            compute_charge_gradients and charge_grads is not None,
-            compute_virial and virial is not None,
-        ):
-            case (True, True, True):
-                return energies, forces, charge_grads, virial
-            case (True, True, False):
-                return energies, forces, charge_grads
-            case (True, False, True):
-                return energies, forces, virial
-            case (True, False, False):
-                return energies, forces
-            case (False, True, True):
-                return energies, charge_grads, virial
-            case (False, True, False):
-                return energies, charge_grads
-            case (False, False, True):
-                return energies, virial
-            case _:
-                return energies
-
     if hybrid_forces:
         pos_d = positions.detach()
         chg_d = charges.detach()
@@ -3049,7 +3051,12 @@ def ewald_reciprocal_space(
                 energies, charges, charge_grads, batch_idx
             )
 
-        return _build_result(energies, forces, charge_grads, virial)
+        return _build_electrostatic_result(
+            ElectrostaticOutputs(energies, forces, charge_grads, virial),
+            compute_forces,
+            compute_charge_gradients,
+            compute_virial,
+        )
 
     if compute_charge_gradients:
         if is_batch:
@@ -3076,7 +3083,12 @@ def ewald_reciprocal_space(
                 )
             )
 
-        return _build_result(energies, forces, charge_grads, virial)
+        return _build_electrostatic_result(
+            ElectrostaticOutputs(energies, forces, charge_grads, virial),
+            compute_forces,
+            compute_charge_gradients,
+            compute_virial,
+        )
 
     # No charge gradients
     if is_batch:
@@ -3090,7 +3102,12 @@ def ewald_reciprocal_space(
                 batch_idx,
                 compute_virial=compute_virial,
             )
-            return _build_result(energies, forces, virial=virial)
+            return _build_electrostatic_result(
+                ElectrostaticOutputs(energies, forces, virial=virial),
+                compute_forces,
+                compute_charge_gradients,
+                compute_virial,
+            )
         else:
             energies, virial = _batch_ewald_reciprocal_space_energy(
                 positions,
@@ -3101,7 +3118,12 @@ def ewald_reciprocal_space(
                 batch_idx,
                 compute_virial=compute_virial,
             )
-            return _build_result(energies, virial=virial)
+            return _build_electrostatic_result(
+                ElectrostaticOutputs(energies, virial=virial),
+                compute_forces,
+                compute_charge_gradients,
+                compute_virial,
+            )
     else:
         if compute_forces:
             energies, forces, virial = _ewald_reciprocal_space_energy_forces(
@@ -3112,7 +3134,12 @@ def ewald_reciprocal_space(
                 alpha,
                 compute_virial=compute_virial,
             )
-            return _build_result(energies, forces, virial=virial)
+            return _build_electrostatic_result(
+                ElectrostaticOutputs(energies, forces, virial=virial),
+                compute_forces,
+                compute_charge_gradients,
+                compute_virial,
+            )
         else:
             energies, virial = _ewald_reciprocal_space_energy(
                 positions,
@@ -3122,465 +3149,12 @@ def ewald_reciprocal_space(
                 alpha,
                 compute_virial=compute_virial,
             )
-            return _build_result(energies, virial=virial)
-
-
-###########################################################################################
-########################### Slab Correction Op ############################################
-###########################################################################################
-
-
-def _prepare_pbc_for_slab(
-    pbc: torch.Tensor | None,
-    num_systems: int,
-    device: torch.device,
-) -> torch.Tensor:
-    """Validate and normalize the pbc tensor for slab correction.
-
-    Parameters
-    ----------
-    pbc : torch.Tensor or None
-        Per-system periodic boundary conditions. Accepted shapes:
-        - None: raises (slab correction requires explicit pbc).
-        - (3,) bool: accepted only for a single system.
-        - (num_systems, 3) bool: used as-is.
-        Batched slab correction requires explicit per-system pbc because slab
-        geometry can differ across systems.
-    num_systems : int
-        Number of systems in the batch.
-    device : torch.device
-        Target device for the output tensor.
-
-    Returns
-    -------
-    torch.Tensor, shape (num_systems, 3), dtype=bool
-        Validated pbc tensor.
-    """
-    if pbc is None:
-        raise ValueError(
-            "slab_correction=True requires an explicit `pbc` argument. "
-            "Pass a (3,) or (B, 3) bool tensor with True for periodic "
-            "directions and False for the non-periodic (slab) direction."
-        )
-    if pbc.dtype != torch.bool:
-        raise ValueError(f"`pbc` must be a bool tensor, got dtype={pbc.dtype}.")
-    if pbc.dim() == 1:
-        if pbc.shape[0] != 3:
-            raise ValueError(
-                f"`pbc` of shape (3,) expected, got shape {tuple(pbc.shape)}."
+            return _build_electrostatic_result(
+                ElectrostaticOutputs(energies, virial=virial),
+                compute_forces,
+                compute_charge_gradients,
+                compute_virial,
             )
-        if num_systems != 1:
-            raise ValueError(
-                "Batched slab correction requires `pbc` shape (B, 3); got "
-                "shape (3,). Pass per-system pbc explicitly."
-            )
-        pbc = pbc.unsqueeze(0).contiguous()
-    elif pbc.dim() == 2:
-        if pbc.shape != (num_systems, 3):
-            raise ValueError(
-                f"`pbc` of shape ({num_systems}, 3) expected, got "
-                f"shape {tuple(pbc.shape)}."
-            )
-    else:
-        raise ValueError(
-            f"`pbc` must be 1D (3,) or 2D (B, 3), got {pbc.dim()}D tensor."
-        )
-    return pbc.to(device=device, dtype=torch.bool)
-
-
-def _slab_energy_output() -> OutputSpec:
-    return OutputSpec("slab_energies", wp.float64, lambda pos, *_: (pos.shape[0],))
-
-
-def _slab_forces_output() -> OutputSpec:
-    return OutputSpec(
-        "slab_forces",
-        lambda pos, *_: get_wp_vec_dtype(pos.dtype),
-        lambda pos, *_: (pos.shape[0], 3),
-    )
-
-
-def _slab_charge_grads_output() -> OutputSpec:
-    return OutputSpec("slab_charge_grads", wp.float64, lambda pos, *_: (pos.shape[0],))
-
-
-def _slab_virial_output() -> OutputSpec:
-    return OutputSpec(
-        "slab_virial",
-        lambda pos, *_: get_wp_mat_dtype(pos.dtype),
-        lambda pos, charges, cell, *_: (cell.shape[0], 3, 3),
-    )
-
-
-def _run_slab_correction_op(
-    positions: torch.Tensor,
-    charges: torch.Tensor,
-    cell: torch.Tensor,
-    pbc: torch.Tensor,
-    batch_idx: torch.Tensor,
-    compute_forces: bool = False,
-    compute_charge_gradients: bool = False,
-    compute_virial: bool = False,
-) -> tuple[torch.Tensor, ...]:
-    """Run the Ewald-style slab custom-op implementation."""
-    num_atoms = positions.shape[0]
-    num_systems = cell.shape[0]
-    device = wp.device_from_torch(positions.device)
-    needs_grad_flag = needs_grad(positions, charges, cell)
-    need_force_kernel = compute_forces or compute_charge_gradients or compute_virial
-    virial_grad = needs_grad_flag and compute_virial
-
-    input_dtype = positions.dtype
-    wp_scalar = get_wp_dtype(input_dtype)
-    wp_vec = get_wp_vec_dtype(input_dtype)
-    wp_mat = get_wp_mat_dtype(input_dtype)
-
-    # Slab geometry is computed inside the kernels so Warp autodiff tracks
-    # cell dependence.
-    wp_positions = warp_from_torch(positions, wp_vec, requires_grad=needs_grad_flag)
-    wp_charges = warp_from_torch(charges, wp_scalar, requires_grad=needs_grad_flag)
-    wp_batch_idx = warp_from_torch(batch_idx, wp.int32)
-    wp_pbc = warp_from_torch(pbc.contiguous(), wp.bool)
-    wp_cell = warp_from_torch(cell, wp_mat, requires_grad=needs_grad_flag)
-
-    # Allocate moment arrays (float64, zero-initialized).
-    # mz and mz2 are (B, 3), with projected M and M2 stored in the
-    # non-periodic pbc-axis slot.
-    mz_t = torch.zeros(num_systems, 3, device=positions.device, dtype=torch.float64)
-    mz2_t = torch.zeros(num_systems, 3, device=positions.device, dtype=torch.float64)
-    qtotal_t = torch.zeros(num_systems, device=positions.device, dtype=torch.float64)
-    wp_mz = warp_from_torch(mz_t, wp.float64, requires_grad=needs_grad_flag)
-    wp_mz2 = warp_from_torch(mz2_t, wp.float64, requires_grad=needs_grad_flag)
-    wp_qtotal = warp_from_torch(qtotal_t, wp.float64, requires_grad=needs_grad_flag)
-
-    # Allocate output arrays
-    energy_in_t = torch.zeros(num_atoms, device=positions.device, dtype=torch.float64)
-    wp_energy_in = warp_from_torch(
-        energy_in_t, wp.float64, requires_grad=needs_grad_flag
-    )
-
-    slab_energies = torch.zeros(num_atoms, device=positions.device, dtype=torch.float64)
-    wp_slab_energies = warp_from_torch(
-        slab_energies, wp.float64, requires_grad=needs_grad_flag
-    )
-
-    output_tensors: list[torch.Tensor] = [slab_energies]
-    backward_kw = {
-        "slab_energies": wp_slab_energies,
-        "positions": wp_positions,
-        "charges": wp_charges,
-        "cell": wp_cell,
-    }
-
-    wp_slab_forces = None
-    if need_force_kernel:
-        slab_forces = torch.zeros(
-            num_atoms, 3, device=positions.device, dtype=input_dtype
-        )
-        wp_slab_forces = warp_from_torch(
-            slab_forces, wp_vec, requires_grad=needs_grad_flag
-        )
-        output_tensors.append(slab_forces)
-        backward_kw["slab_forces"] = wp_slab_forces
-
-    wp_slab_charge_grads = None
-    if compute_charge_gradients:
-        slab_charge_grads = torch.zeros(
-            num_atoms, device=positions.device, dtype=torch.float64
-        )
-        wp_slab_charge_grads = warp_from_torch(
-            slab_charge_grads, wp.float64, requires_grad=needs_grad_flag
-        )
-        output_tensors.append(slab_charge_grads)
-        backward_kw["slab_charge_grads"] = wp_slab_charge_grads
-
-    wp_slab_virial = None
-    if need_force_kernel:
-        slab_virial = torch.zeros(
-            num_systems, 3, 3, device=positions.device, dtype=input_dtype
-        )
-        wp_slab_virial = warp_from_torch(slab_virial, wp_mat, requires_grad=virial_grad)
-        output_tensors.append(slab_virial)
-        if virial_grad:
-            backward_kw["slab_virial"] = wp_slab_virial
-
-    with WarpAutogradContextManager(needs_grad_flag) as tape:
-        wp.launch(
-            _slab_reduce_moments_kernel_overload[wp_scalar],
-            dim=num_atoms,
-            inputs=[
-                wp_positions,
-                wp_charges,
-                wp_batch_idx,
-                wp_pbc,
-                wp_cell,
-                wp_mz,
-                wp_mz2,
-                wp_qtotal,
-            ],
-            device=device,
-        )
-
-        _launch_slab_correction(
-            positions=wp_positions,
-            charges=wp_charges,
-            batch_idx=wp_batch_idx,
-            pbc=wp_pbc,
-            cell=wp_cell,
-            mz=wp_mz,
-            mz2=wp_mz2,
-            qtotal=wp_qtotal,
-            energy_in=wp_energy_in,
-            energy_out=wp_slab_energies,
-            wp_dtype=wp_scalar,
-            forces=wp_slab_forces,
-            charge_grads=wp_slab_charge_grads,
-            virial=wp_slab_virial,
-            compute_forces=compute_forces,
-            compute_charge_gradients=compute_charge_gradients,
-            compute_virial=compute_virial,
-            device=device,
-        )
-
-    if needs_grad_flag:
-        attach_for_backward(slab_energies, tape=tape, **backward_kw)
-        stable_outputs = [slab_energies]
-        stable_outputs.extend(tensor.clone() for tensor in output_tensors[1:])
-        return tuple(stable_outputs)
-
-    return tuple(output_tensors)
-
-
-@warp_custom_op(
-    name="alchemiops::_slab_correction_energy",
-    outputs=[_slab_energy_output()],
-    grad_arrays=["slab_energies", "positions", "charges", "cell"],
-)
-def _slab_correction_energy_op(
-    positions: torch.Tensor,
-    charges: torch.Tensor,
-    cell: torch.Tensor,
-    pbc: torch.Tensor,
-    batch_idx: torch.Tensor,
-) -> torch.Tensor:
-    """Compute slab correction energy only."""
-    return _run_slab_correction_op(positions, charges, cell, pbc, batch_idx)[0]
-
-
-@warp_custom_op(
-    name="alchemiops::_slab_correction_energy_forces",
-    outputs=[_slab_energy_output(), _slab_forces_output(), _slab_virial_output()],
-    grad_arrays=[
-        "slab_energies",
-        "slab_forces",
-        "slab_virial",
-        "positions",
-        "charges",
-        "cell",
-    ],
-)
-def _slab_correction_energy_forces_op(
-    positions: torch.Tensor,
-    charges: torch.Tensor,
-    cell: torch.Tensor,
-    pbc: torch.Tensor,
-    batch_idx: torch.Tensor,
-    compute_virial: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """Compute slab correction energy, forces, and optionally virial."""
-    return _run_slab_correction_op(
-        positions,
-        charges,
-        cell,
-        pbc,
-        batch_idx,
-        compute_forces=True,
-        compute_virial=compute_virial,
-    )
-
-
-@warp_custom_op(
-    name="alchemiops::_slab_correction_energy_forces_charge_grad",
-    outputs=[
-        _slab_energy_output(),
-        _slab_forces_output(),
-        _slab_charge_grads_output(),
-        _slab_virial_output(),
-    ],
-    grad_arrays=[
-        "slab_energies",
-        "slab_forces",
-        "slab_charge_grads",
-        "slab_virial",
-        "positions",
-        "charges",
-        "cell",
-    ],
-)
-def _slab_correction_energy_forces_charge_grad_op(
-    positions: torch.Tensor,
-    charges: torch.Tensor,
-    cell: torch.Tensor,
-    pbc: torch.Tensor,
-    batch_idx: torch.Tensor,
-    compute_virial: bool = False,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
-    """Compute slab correction energy, forces, charge gradients, and optionally virial."""
-    return _run_slab_correction_op(
-        positions,
-        charges,
-        cell,
-        pbc,
-        batch_idx,
-        compute_forces=True,
-        compute_charge_gradients=True,
-        compute_virial=compute_virial,
-    )
-
-
-def compute_slab_correction(
-    positions: torch.Tensor,
-    charges: torch.Tensor,
-    cell: torch.Tensor,
-    pbc: torch.Tensor,
-    batch_idx: torch.Tensor | None = None,
-    compute_forces: bool = False,
-    compute_charge_gradients: bool = False,
-    compute_virial: bool = False,
-) -> torch.Tensor | tuple[torch.Tensor, ...]:
-    """Yeh-Berkowitz slab correction for 2D periodic electrostatics, with the
-    Ballenegger et al. (2009) Eq. 29 extension for non-neutral systems.
-
-    Returns the slab-correction contribution (per-atom energy and optionally
-    per-atom force, charge gradient, and per-system virial). The caller adds
-    these to the corresponding 3D Ewald quantities; in normal usage the
-    correction is invoked through ``ewald_summation(..., slab_correction=True)``.
-
-    Background-charge convention
-    ----------------------------
-    For systems with net charge :math:`Q \\ne 0`, the formula corresponds to a
-    uniform-volume neutralizing background (the same convention used by
-    standard 3D Ewald). This matches LAMMPS and Ballenegger et al. (2009)
-    Eq. 29.
-
-    Cell geometry
-    -------------
-    Orthorhombic and triclinic cells are supported. For triclinic slab
-    systems, the slab normal follows the plane spanned by the two periodic
-    cell vectors.
-
-    Parameters
-    ----------
-    positions : torch.Tensor, shape (N, 3)
-        Atomic coordinates.
-    charges : torch.Tensor, shape (N,)
-        Atomic charges.
-    cell : torch.Tensor, shape (3, 3) or (B, 3, 3)
-        Unit cell matrices.
-    pbc : torch.Tensor, shape (3,) or (B, 3), dtype=bool
-        Per-system periodic boundary conditions. True for periodic directions,
-        False for the non-periodic (slab) direction. Systems whose pbc is not
-        slab-like (i.e., has anything other than exactly one False entry)
-        contribute zero. A (3,) tensor is accepted only for single-system
-        calls; batched calls require explicit (B, 3) per-system pbc.
-    batch_idx : torch.Tensor, shape (N,), dtype=int32, optional
-        System index for each atom. Defaults to all zeros (single system).
-    compute_forces : bool, default=False
-        If True, return per-atom forces.
-    compute_charge_gradients : bool, default=False
-        If True, return per-atom charge gradients dE_slab/dq_i.
-    compute_virial : bool, default=False
-        If True, return per-system virial tensor using the normal-following
-        affine strain convention W = E_slab * (I - 2 n n^T).
-
-    Returns
-    -------
-    energies : torch.Tensor, shape (N,), dtype=float64
-        Per-atom slab correction energy.
-    forces : torch.Tensor, shape (N, 3), dtype matches positions, optional
-        Per-atom slab force (only returned if compute_forces=True).
-    charge_grads : torch.Tensor, shape (N,), dtype=float64, optional
-        Per-atom slab charge gradient (only if compute_charge_gradients=True).
-    virial : torch.Tensor, shape (B, 3, 3), dtype matches positions, optional
-        Per-system slab virial (only if compute_virial=True).
-
-    Examples
-    --------
-    Standalone correction for an orthorhombic slab with vacuum along z::
-
-        >>> pbc_slab = torch.tensor([[True, True, False]], device=positions.device)
-        >>> slab_energy, slab_forces = compute_slab_correction(
-        ...     positions, charges, cell, pbc_slab, compute_forces=True
-        ... )
-        >>> corrected_energy = ewald_energy + slab_energy
-
-    Triclinic cells use the normal to the periodic plane::
-
-        >>> triclinic_energy, triclinic_forces = compute_slab_correction(
-        ...     positions, charges, triclinic_cell, pbc_slab, compute_forces=True
-        ... )
-    """
-    cell, num_systems = _prepare_cell(cell)
-    pbc = _prepare_pbc_for_slab(pbc, num_systems, positions.device)
-
-    if batch_idx is None:
-        batch_idx = torch.zeros(
-            positions.shape[0], dtype=torch.int32, device=positions.device
-        )
-
-    def _build_result(
-        energies: torch.Tensor,
-        forces: torch.Tensor | None = None,
-        charge_grads: torch.Tensor | None = None,
-        virial: torch.Tensor | None = None,
-    ) -> torch.Tensor | tuple[torch.Tensor, ...]:
-        match (
-            compute_forces and forces is not None,
-            compute_charge_gradients and charge_grads is not None,
-            compute_virial and virial is not None,
-        ):
-            case (True, True, True):
-                return energies, forces, charge_grads, virial
-            case (True, True, False):
-                return energies, forces, charge_grads
-            case (True, False, True):
-                return energies, forces, virial
-            case (True, False, False):
-                return energies, forces
-            case (False, True, True):
-                return energies, charge_grads, virial
-            case (False, True, False):
-                return energies, charge_grads
-            case (False, False, True):
-                return energies, virial
-            case _:
-                return energies
-
-    if compute_charge_gradients:
-        energies, forces, charge_grads, virial = (
-            _slab_correction_energy_forces_charge_grad_op(
-                positions,
-                charges,
-                cell,
-                pbc,
-                batch_idx,
-                compute_virial=compute_virial,
-            )
-        )
-        return _build_result(energies, forces, charge_grads, virial)
-
-    if compute_forces or compute_virial:
-        energies, forces, virial = _slab_correction_energy_forces_op(
-            positions,
-            charges,
-            cell,
-            pbc,
-            batch_idx,
-            compute_virial=compute_virial,
-        )
-        return _build_result(energies, forces, virial=virial)
-
-    return _slab_correction_energy_op(positions, charges, cell, pbc, batch_idx)
 
 
 def ewald_summation(
@@ -3774,11 +3348,11 @@ def ewald_summation(
     )
 
     # Optional slab correction: returns a same-shape tuple as rs/rec,
-    # so it composes uniformly with the element-wise combination below.
-    slab_tuple: tuple[torch.Tensor, ...] | None = None
+    # so it composes uniformly with the named-field combination below.
+    slab_result: torch.Tensor | tuple[torch.Tensor, ...] | None = None
     if slab_correction:
         if hybrid_forces:
-            slab_out = compute_slab_correction(
+            slab_out = _compute_slab_correction(
                 positions.detach(),
                 charges.detach(),
                 cell.detach(),
@@ -3788,35 +3362,32 @@ def ewald_summation(
                 compute_charge_gradients=True,
                 compute_virial=compute_virial,
             )
-            slab_raw = slab_out if isinstance(slab_out, tuple) else (slab_out,)
-
-            if compute_forces and compute_virial:
-                slab_energies, slab_forces, slab_charge_grads, slab_virial = slab_raw
-            elif compute_forces:
-                slab_energies, slab_forces, slab_charge_grads = slab_raw
-                slab_virial = None
-            elif compute_virial:
-                slab_energies, slab_charge_grads, slab_virial = slab_raw
-                slab_forces = None
-            else:
-                slab_energies, slab_charge_grads = slab_raw
-                slab_forces = None
-                slab_virial = None
+            slab = _unpack_electrostatic_outputs(
+                slab_out,
+                compute_forces,
+                compute_charge_gradients=True,
+                compute_virial=compute_virial,
+            )
+            slab_energies = slab.energies
 
             if charges.requires_grad:
                 slab_energies = _InjectChargeGrad.apply(
-                    slab_energies, charges, slab_charge_grads, batch_idx
+                    slab_energies, charges, slab.charge_grads, batch_idx
                 )
 
-            slab_tuple = (slab_energies,)
-            if compute_forces and slab_forces is not None:
-                slab_tuple += (slab_forces,)
-            if compute_charge_gradients:
-                slab_tuple += (slab_charge_grads,)
-            if compute_virial and slab_virial is not None:
-                slab_tuple += (slab_virial,)
+            slab_result = _build_electrostatic_result(
+                ElectrostaticOutputs(
+                    slab_energies,
+                    slab.forces,
+                    slab.charge_grads,
+                    slab.virial,
+                ),
+                compute_forces,
+                compute_charge_gradients,
+                compute_virial,
+            )
         else:
-            slab_out = compute_slab_correction(
+            slab_result = _compute_slab_correction(
                 positions,
                 charges,
                 cell,
@@ -3826,46 +3397,12 @@ def ewald_summation(
                 compute_charge_gradients=compute_charge_gradients,
                 compute_virial=compute_virial,
             )
-            slab_tuple = slab_out if isinstance(slab_out, tuple) else (slab_out,)
 
-    # Normalize return tuples for element-wise combination
-    rs_tuple = rs if isinstance(rs, tuple) else (rs,)
-    rec_tuple = rec if isinstance(rec, tuple) else (rec,)
-    tuple_index = 0
-
-    energies = rs_tuple[tuple_index] + rec_tuple[tuple_index]
-    if slab_tuple is not None:
-        energies = energies + slab_tuple[tuple_index]
-    tuple_index += 1
-    results: tuple[torch.Tensor, ...] = (energies,)
-
-    if compute_forces:
-        forces = rs_tuple[tuple_index] + rec_tuple[tuple_index]
-        if slab_tuple is not None:
-            forces = forces + slab_tuple[tuple_index]
-        results += (forces,)
-        tuple_index += 1
-
-    if compute_charge_gradients:
-        real_charge_grads = rs_tuple[tuple_index]
-        reciprocal_charge_grads = rec_tuple[tuple_index]
-        if torch.compiler.is_compiling():
-            charge_grads = _sum_charge_gradients(
-                real_charge_grads, reciprocal_charge_grads
-            )
-        else:
-            charge_grads = real_charge_grads + reciprocal_charge_grads
-        if slab_tuple is not None:
-            charge_grads = charge_grads + slab_tuple[tuple_index]
-        results += (charge_grads,)
-        tuple_index += 1
-
-    if compute_virial:
-        virial = rs_tuple[tuple_index] + rec_tuple[tuple_index]
-        if slab_tuple is not None:
-            virial = virial + slab_tuple[tuple_index]
-        results += (virial,)
-
-    if len(results) == 1:
-        return results[0]
-    return results
+    return _combine_electrostatic_outputs(
+        rs,
+        rec,
+        slab_result,
+        compute_forces,
+        compute_charge_gradients,
+        compute_virial,
+    )
